Restructure Effects, no code was changed.
