+++ src/synthesizer/audio_engine/effects/gs/insertion/convert.ts
