+++ src/synthesizer/audio_engine/effects/gs/insertion/stereo_eq.ts
@@ -1,4 +1,4 @@
-import type { InsertionProcessor } from "../types";
+import type { InsertionProcessor } from "../../types";
 import { InsertionValueConverter } from "./convert";
 import {
     type BiquadCoeffs,
