+++ src/synthesizer/audio_engine/effects/gs/insertion_list.ts
@@ -1,4 +1,4 @@
-import type { InsertionProcessorConstructor } from "./types";
+import type { InsertionProcessorConstructor } from "../types";
 import { ThruFX } from "./insertion/thru";
 import { StereoEQFX } from "./insertion/stereo_eq";
 import { PhaserFX } from "./insertion/phaser";
